Move BaseWidget to separate file
